+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
   - Enums config class to change the default policy for binding and mapping Enum values.
   - @UseEnumStrategy SqlObject configurer for setting the default enum strategy.
   - @EnumByName and @EnumByOrdinal qualifying annotations to override said default policy.
+  - Support for Postgres custom types, both user defined and driver types like `PGInterval` or `PGcircle`
 - Improvements
   - Improve Error handling with transactions
 - Minor source incompatibility